dummy_docker/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
tensorflow-cpu==2.15.0
numpy==1.24.2